Rename "secure browsers" to "browsers with countermeasures" (#111)
This PR renames the section "secure browsers" to "browsers with
countermeasures" due to the fact that the entry criterion to the list
is not the capability of thwarting nothingprivate.ml, but instead
thwarting nothingprivate.ml in a specific way, namely randomizing the
canvas fingerprint, which can open up new fingerprinting opportunities.
Therefore implying that the list of browsers is secure while deliberately
excluding browsers that are equally capable of thwarting nothingprivate.ml
by removing the canvas fingerprint. Furthermore "secure browsers"
implies that these browsers are capable of fully protecting against
fingerprinting while being tested only against a very specific
fingerprinting routine.

Therefore, to avoid implying misinformation, it this PR replaces "secure
browsers" with a more specific description to not imply misinformation.